Add typeAnnotation to ObjectPattern in flow
When investigating https://github.com/prettier/prettier/issues/1007, it turns out that `types.getFieldNames(node)` for a node of type `ObjectPattern` doesn't return `typeAnnotation` and therefore it doesn't show up in the traversal for comment location.

I've patched this in my `node_modules/` folder and it fixes the comment location. Let me know if you need anything else to ensure that this is good.