Don't worry about very early stage proposals.
Whenever Babel adds support for an experimental proposal to @babel/parser
and @babel/types, the ast-types completeness tests fail.

This commit allows us to wait until the proposal has advanced to a later
stage (let's say >1, at the very least).

Besides the risk that the syntax will never become part of the ECMAScript
standard, if it's only at Stage 1 then its AST format will almost
certainly change before the spec is finalized, which leads to subtle
breaking changes for users of ast-types. Better to wait to implement the
type until it's more stable.