Make 'optional' field of ChainElement false by default.
Since CallExpression and MemberExpression inherit from this type, and
those types are not optional by default, the default value of the
'optional' field must be false.

As a side benefit, I think this default gives more value to the (otherwise
somewhat redundant) OptionalMemberExpression and OptionalCallExpression
types, whose 'optional' fields are true by default, in contrast to the
ordinary MemberExpression and CallExpression types.