Install recast@next instead of temporary @pr-1247 version.
I am intentionally using the ^0.23.0 syntax here, even though other
versions in devDependencies are exact (no ^), because recast has an
exact dependency on ast-types. If there was an exact dependency in the
other direction too, and the versions ever disagreed, multiple copies of
these packages would need to be installed in nested node_modules
directories, which is undesirable for many reasons.